Refactore the code structure
