+++ src/App.js
@@ -1,5 +1,5 @@
 import Navbar from './components/Navbar';
-import TheRouter from './components/TheRouters';
+import TheRouter from './pages/TheRouters';
 
 export default function App() {
   return (
